requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
PySocks==1.7.0
PyYAML==5.1.1
requests==2.22.0
urllib3==1.25.3
